WIP: Refactoring receipt handling to be done in VM logic
