+++ runtime/near-vm-logic/src/dependencies.rs
@@ -1,9 +1,6 @@
 //! External dependencies of the near-vm-logic.
 
-use crate::types::{PublicKey, ReceiptIndex};
-use near_primitives_core::types::{AccountId, Balance, Gas};
-#[cfg(feature = "protocol_feature_function_call_weight")]
-use near_primitives_core::types::{GasDistribution, GasWeight};
+use near_primitives_core::types::{AccountId, Balance};
 use near_vm_errors::VMLogicError;
 
 /// An abstraction over the memory of the smart contract.
@@ -164,371 +161,371 @@ pub trait External {
     /// ```
     fn storage_has_key(&mut self, key: &[u8]) -> Result<bool>;
 
-    /// Create a receipt which will be executed after all the receipts identified by
-    /// `receipt_indices` are complete.
-    ///
-    /// If any of the [`RecepitIndex`]es do not refer to a known receipt, this function will fail
-    /// with an error.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_indices` - a list of receipt indices the new receipt is depend on
-    ///
-    /// # Example
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index_one = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// let receipt_index_two = external.create_receipt(vec![receipt_index_one], "bob.near".parse().unwrap());
-    ///
-    /// ```
-    fn create_receipt(
-        &mut self,
-        receipt_indices: Vec<ReceiptIndex>,
-        receiver_id: AccountId,
-    ) -> Result<ReceiptIndex>;
+    // /// Create a receipt which will be executed after all the receipts identified by
+    // /// `receipt_indices` are complete.
+    // ///
+    // /// If any of the [`RecepitIndex`]es do not refer to a known receipt, this function will fail
+    // /// with an error.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_indices` - a list of receipt indices the new receipt is depend on
+    // ///
+    // /// # Example
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index_one = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// let receipt_index_two = external.create_receipt(vec![receipt_index_one], "bob.near".parse().unwrap());
+    // ///
+    // /// ```
+    // fn create_receipt(
+    //     &mut self,
+    //     receipt_indices: Vec<ReceiptIndex>,
+    //     receiver_id: AccountId,
+    // ) -> Result<ReceiptIndex>;
 
-    /// Attach the [`CreateAccountAction`] action to an existing receipt.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    ///
-    /// # Example
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_create_account(receipt_index).unwrap();
-    ///
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_create_account(&mut self, receipt_index: ReceiptIndex) -> Result<()>;
+    // /// Attach the [`CreateAccountAction`] action to an existing receipt.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // ///
+    // /// # Example
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_create_account(receipt_index).unwrap();
+    // ///
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_create_account(&mut self, receipt_index: ReceiptIndex) -> Result<()>;
 
-    /// Attach the [`DeployContractAction`] action to an existing receipt.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `code` - a Wasm code to attach
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_deploy_contract(receipt_index, b"some valid Wasm code".to_vec()).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_deploy_contract(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        code: Vec<u8>,
-    ) -> Result<()>;
+    // /// Attach the [`DeployContractAction`] action to an existing receipt.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `code` - a Wasm code to attach
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_deploy_contract(receipt_index, b"some valid Wasm code".to_vec()).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_deploy_contract(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     code: Vec<u8>,
+    // ) -> Result<()>;
 
-    /// Attach the [`FunctionCallAction`] action to an existing receipt.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `method_name` - a name of the contract method to call
-    /// * `arguments` - a Wasm code to attach
-    /// * `attached_deposit` - amount of tokens to transfer with the call
-    /// * `prepaid_gas` - amount of prepaid gas to attach to the call
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_function_call(
-    ///     receipt_index,
-    ///     b"method_name".to_vec(),
-    ///     b"{serialised: arguments}".to_vec(),
-    ///     100000u128,
-    ///     100u64
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_function_call(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        method_name: Vec<u8>,
-        arguments: Vec<u8>,
-        attached_deposit: Balance,
-        prepaid_gas: Gas,
-    ) -> Result<()>;
+    // /// Attach the [`FunctionCallAction`] action to an existing receipt.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `method_name` - a name of the contract method to call
+    // /// * `arguments` - a Wasm code to attach
+    // /// * `attached_deposit` - amount of tokens to transfer with the call
+    // /// * `prepaid_gas` - amount of prepaid gas to attach to the call
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_function_call(
+    // ///     receipt_index,
+    // ///     b"method_name".to_vec(),
+    // ///     b"{serialised: arguments}".to_vec(),
+    // ///     100000u128,
+    // ///     100u64
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_function_call(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     method_name: Vec<u8>,
+    //     arguments: Vec<u8>,
+    //     attached_deposit: Balance,
+    //     prepaid_gas: Gas,
+    // ) -> Result<()>;
 
-    /// Attach the [`FunctionCallAction`] action to an existing receipt. This method has similar
-    /// functionality to [`append_action_function_call`](Self::append_action_function_call) except
-    /// that it allows specifying a weight to use leftover gas from the current execution.
-    ///
-    /// `prepaid_gas` and `gas_weight` can either be specified or both. If a `gas_weight` is
-    /// specified, the action should be allocated gas in
-    /// [`distribute_unused_gas`](Self::distribute_unused_gas).
-    ///
-    /// For more information, see [crate::VMLogic::promise_batch_action_function_call_weight].
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `method_name` - a name of the contract method to call
-    /// * `arguments` - a Wasm code to attach
-    /// * `attached_deposit` - amount of tokens to transfer with the call
-    /// * `prepaid_gas` - amount of prepaid gas to attach to the call
-    /// * `gas_weight` - relative weight of unused gas to distribute to the function call action
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_function_call_weight(
-    ///     receipt_index,
-    ///     b"method_name".to_vec(),
-    ///     b"{serialised: arguments}".to_vec(),
-    ///     100000u128,
-    ///     100u64,
-    ///     2,
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    #[cfg(feature = "protocol_feature_function_call_weight")]
-    fn append_action_function_call_weight(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        method_name: Vec<u8>,
-        arguments: Vec<u8>,
-        attached_deposit: Balance,
-        prepaid_gas: Gas,
-        gas_weight: GasWeight,
-    ) -> Result<()>;
+    // /// Attach the [`FunctionCallAction`] action to an existing receipt. This method has similar
+    // /// functionality to [`append_action_function_call`](Self::append_action_function_call) except
+    // /// that it allows specifying a weight to use leftover gas from the current execution.
+    // ///
+    // /// `prepaid_gas` and `gas_weight` can either be specified or both. If a `gas_weight` is
+    // /// specified, the action should be allocated gas in
+    // /// [`distribute_unused_gas`](Self::distribute_unused_gas).
+    // ///
+    // /// For more information, see [crate::VMLogic::promise_batch_action_function_call_weight].
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `method_name` - a name of the contract method to call
+    // /// * `arguments` - a Wasm code to attach
+    // /// * `attached_deposit` - amount of tokens to transfer with the call
+    // /// * `prepaid_gas` - amount of prepaid gas to attach to the call
+    // /// * `gas_weight` - relative weight of unused gas to distribute to the function call action
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_function_call_weight(
+    // ///     receipt_index,
+    // ///     b"method_name".to_vec(),
+    // ///     b"{serialised: arguments}".to_vec(),
+    // ///     100000u128,
+    // ///     100u64,
+    // ///     2,
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // #[cfg(feature = "protocol_feature_function_call_weight")]
+    // fn append_action_function_call_weight(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     method_name: Vec<u8>,
+    //     arguments: Vec<u8>,
+    //     attached_deposit: Balance,
+    //     prepaid_gas: Gas,
+    //     gas_weight: GasWeight,
+    // ) -> Result<()>;
 
-    /// Attach the [`TransferAction`] action to an existing receipt.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `amount` - amount of tokens to transfer
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_transfer(
-    ///     receipt_index,
-    ///     100000u128,
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_transfer(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        amount: Balance,
-    ) -> Result<()>;
+    // /// Attach the [`TransferAction`] action to an existing receipt.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `amount` - amount of tokens to transfer
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_transfer(
+    // ///     receipt_index,
+    // ///     100000u128,
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_transfer(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     amount: Balance,
+    // ) -> Result<()>;
 
-    /// Attach the [`StakeAction`] action to an existing receipt.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `stake` - amount of tokens to stake
-    /// * `public_key` - a validator public key
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_stake(
-    ///     receipt_index,
-    ///     100000u128,
-    ///     b"some public key".to_vec()
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_stake(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        stake: Balance,
-        public_key: PublicKey,
-    ) -> Result<()>;
+    // /// Attach the [`StakeAction`] action to an existing receipt.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `stake` - amount of tokens to stake
+    // /// * `public_key` - a validator public key
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_stake(
+    // ///     receipt_index,
+    // ///     100000u128,
+    // ///     b"some public key".to_vec()
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_stake(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     stake: Balance,
+    //     public_key: PublicKey,
+    // ) -> Result<()>;
 
-    /// Attach the [`AddKeyAction`] action to an existing receipt.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `public_key` - a public key for an access key
-    /// * `nonce` - a nonce
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_add_key_with_full_access(
-    ///     receipt_index,
-    ///     b"some public key".to_vec(),
-    ///     0u64
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_add_key_with_full_access(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        public_key: PublicKey,
-        nonce: u64,
-    ) -> Result<()>;
+    // /// Attach the [`AddKeyAction`] action to an existing receipt.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `public_key` - a public key for an access key
+    // /// * `nonce` - a nonce
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_add_key_with_full_access(
+    // ///     receipt_index,
+    // ///     b"some public key".to_vec(),
+    // ///     0u64
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_add_key_with_full_access(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     public_key: PublicKey,
+    //     nonce: u64,
+    // ) -> Result<()>;
 
-    /// Attach the [`AddKeyAction`] action an existing receipt.
-    ///
-    /// The access key associated with the action will have the
-    /// [`AccessKeyPermission::FunctionCall`] permission scope.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `public_key` - a public key for an access key
-    /// * `nonce` - a nonce
-    /// * `allowance` - amount of tokens allowed to spend by this access key
-    /// * `receiver_id` - a contract witch will be allowed to call with this access key
-    /// * `method_names` - a list of method names is allowed to call with this access key (empty = any method)
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_add_key_with_function_call(
-    ///     receipt_index,
-    ///     b"some public key".to_vec(),
-    ///     0u64,
-    ///     None,
-    ///     "bob.near".parse().unwrap(),
-    ///     vec![b"foo".to_vec(), b"bar".to_vec()]
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_add_key_with_function_call(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        public_key: PublicKey,
-        nonce: u64,
-        allowance: Option<Balance>,
-        receiver_id: AccountId,
-        method_names: Vec<Vec<u8>>,
-    ) -> Result<()>;
+    // /// Attach the [`AddKeyAction`] action an existing receipt.
+    // ///
+    // /// The access key associated with the action will have the
+    // /// [`AccessKeyPermission::FunctionCall`] permission scope.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `public_key` - a public key for an access key
+    // /// * `nonce` - a nonce
+    // /// * `allowance` - amount of tokens allowed to spend by this access key
+    // /// * `receiver_id` - a contract witch will be allowed to call with this access key
+    // /// * `method_names` - a list of method names is allowed to call with this access key (empty = any method)
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_add_key_with_function_call(
+    // ///     receipt_index,
+    // ///     b"some public key".to_vec(),
+    // ///     0u64,
+    // ///     None,
+    // ///     "bob.near".parse().unwrap(),
+    // ///     vec![b"foo".to_vec(), b"bar".to_vec()]
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_add_key_with_function_call(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     public_key: PublicKey,
+    //     nonce: u64,
+    //     allowance: Option<Balance>,
+    //     receiver_id: AccountId,
+    //     method_names: Vec<Vec<u8>>,
+    // ) -> Result<()>;
 
-    /// Attach the [`DeleteKeyAction`] action to an existing receipt.
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `public_key` - a public key for an access key to delete
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_delete_key(
-    ///     receipt_index,
-    ///     b"some public key".to_vec()
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_delete_key(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        public_key: PublicKey,
-    ) -> Result<()>;
+    // /// Attach the [`DeleteKeyAction`] action to an existing receipt.
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `public_key` - a public key for an access key to delete
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_delete_key(
+    // ///     receipt_index,
+    // ///     b"some public key".to_vec()
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_delete_key(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     public_key: PublicKey,
+    // ) -> Result<()>;
 
-    /// Attach the [`DeleteAccountAction`] action to an existing receipt
-    ///
-    /// # Arguments
-    ///
-    /// * `receipt_index` - an index of Receipt to append an action
-    /// * `beneficiary_id` - an account id to which the rest of the funds of the removed account will be transferred
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use near_vm_logic::mocks::mock_external::MockedExternal;
-    /// # use near_vm_logic::External;
-    ///
-    /// # let mut external = MockedExternal::new();
-    /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
-    /// external.append_action_delete_account(
-    ///     receipt_index,
-    ///     "sam".parse().unwrap()
-    /// ).unwrap();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if the `receipt_index` does not refer to a known receipt.
-    fn append_action_delete_account(
-        &mut self,
-        receipt_index: ReceiptIndex,
-        beneficiary_id: AccountId,
-    ) -> Result<()>;
+    // /// Attach the [`DeleteAccountAction`] action to an existing receipt
+    // ///
+    // /// # Arguments
+    // ///
+    // /// * `receipt_index` - an index of Receipt to append an action
+    // /// * `beneficiary_id` - an account id to which the rest of the funds of the removed account will be transferred
+    // ///
+    // /// # Example
+    // ///
+    // /// ```
+    // /// # use near_vm_logic::mocks::mock_external::MockedExternal;
+    // /// # use near_vm_logic::External;
+    // ///
+    // /// # let mut external = MockedExternal::new();
+    // /// let receipt_index = external.create_receipt(vec![], "charli.near".parse().unwrap()).unwrap();
+    // /// external.append_action_delete_account(
+    // ///     receipt_index,
+    // ///     "sam".parse().unwrap()
+    // /// ).unwrap();
+    // /// ```
+    // ///
+    // /// # Panics
+    // ///
+    // /// Panics if the `receipt_index` does not refer to a known receipt.
+    // fn append_action_delete_account(
+    //     &mut self,
+    //     receipt_index: ReceiptIndex,
+    //     beneficiary_id: AccountId,
+    // ) -> Result<()>;
 
     /// Returns amount of touched trie nodes by storage operations
     fn get_touched_nodes_count(&self) -> u64;
@@ -540,15 +537,6 @@ pub trait External {
     /// Returns total stake of validators in the current epoch.
     fn validator_total_stake(&self) -> Result<Balance>;
 
-    /// Distribute the gas among the scheduled function calls that specify a gas weight.
-    ///
-    /// # Arguments
-    ///
-    /// * `gas` - amount of unused gas to distribute
-    ///
-    /// # Returns
-    ///
-    /// Function returns a [GasDistribution] that indicates how the gas was distributed.
-    #[cfg(feature = "protocol_feature_function_call_weight")]
-    fn distribute_unused_gas(&mut self, gas: Gas) -> GasDistribution;
+    // #[cfg(feature = "protocol_feature_function_call_weight")]
+    // fn distribute_unused_gas(&mut self, gas: Gas) -> GasDistribution;
 }
