+++ runtime/runtime/src/actions.rs
@@ -140,9 +140,6 @@ pub(crate) fn action_function_call(
     let mut runtime_ext = RuntimeExt::new(
         state_update,
         account_id,
-        &action_receipt.signer_id,
-        &action_receipt.signer_public_key,
-        action_receipt.gas_price,
         action_hash,
         &apply_state.epoch_id,
         &apply_state.prev_block_hash,
@@ -239,7 +236,12 @@ pub(crate) fn action_function_call(
             account.set_amount(outcome.balance);
             account.set_storage_usage(outcome.storage_usage);
             result.result = Ok(outcome.return_data);
-            result.new_receipts.extend(runtime_ext.into_receipts(account_id));
+            result.new_receipts.extend(outcome.receipt_manager.into_receipts(
+                &receipt.predecessor_id,
+                &action_receipt.signer_id,
+                &action_receipt.signer_public_key,
+                action_receipt.gas_price,
+            ));
         }
     } else {
         assert!(!execution_succeeded, "Outcome should always be available if execution succeeded")
