+++ runtime/runtime/src/ext.rs
@@ -1,35 +1,21 @@
 use std::sync::Arc;
 
-use borsh::BorshDeserialize;
 use tracing::debug;
 
-use near_crypto::PublicKey;
-use near_primitives::account::{AccessKey, AccessKeyPermission, FunctionCallPermission};
 use near_primitives::contract::ContractCode;
 use near_primitives::errors::{EpochError, StorageError};
 use near_primitives::hash::CryptoHash;
-use near_primitives::receipt::{ActionReceipt, DataReceiver, Receipt, ReceiptEnum};
-use near_primitives::transaction::{
-    Action, AddKeyAction, CreateAccountAction, DeleteAccountAction, DeleteKeyAction,
-    DeployContractAction, FunctionCallAction, StakeAction, TransferAction,
-};
 use near_primitives::trie_key::{trie_key_parsers, TrieKey};
-use near_primitives::types::{AccountId, Balance, EpochId, EpochInfoProvider, Gas, TrieCacheMode};
-#[cfg(feature = "protocol_feature_function_call_weight")]
-use near_primitives::types::{GasDistribution, GasWeight};
+use near_primitives::types::{AccountId, Balance, EpochId, EpochInfoProvider, TrieCacheMode};
 use near_primitives::utils::create_data_id;
 use near_primitives::version::ProtocolVersion;
 use near_store::{get_code, TrieUpdate, TrieUpdateValuePtr};
-use near_vm_errors::{AnyError, HostError, VMLogicError};
+use near_vm_errors::{AnyError, VMLogicError};
 use near_vm_logic::{External, ValuePtr};
 
 pub struct RuntimeExt<'a> {
     trie_update: &'a mut TrieUpdate,
     account_id: &'a AccountId,
-    action_receipts: Vec<(AccountId, ActionReceipt)>,
-    signer_id: &'a AccountId,
-    signer_public_key: &'a PublicKey,
-    gas_price: Balance,
     action_hash: &'a CryptoHash,
     data_count: u64,
     epoch_id: &'a EpochId,
@@ -37,15 +23,6 @@ pub struct RuntimeExt<'a> {
     last_block_hash: &'a CryptoHash,
     epoch_info_provider: &'a dyn EpochInfoProvider,
     current_protocol_version: ProtocolVersion,
-
-    #[cfg(feature = "protocol_feature_function_call_weight")]
-    gas_weights: Vec<(FunctionCallActionIndex, GasWeight)>,
-}
-
-#[cfg(feature = "protocol_feature_function_call_weight")]
-struct FunctionCallActionIndex {
-    receipt_index: usize,
-    action_index: usize,
 }
 
 /// Error used by `RuntimeExt`.
@@ -80,9 +57,6 @@ impl<'a> RuntimeExt<'a> {
     pub fn new(
         trie_update: &'a mut TrieUpdate,
         account_id: &'a AccountId,
-        signer_id: &'a AccountId,
-        signer_public_key: &'a PublicKey,
-        gas_price: Balance,
         action_hash: &'a CryptoHash,
         epoch_id: &'a EpochId,
         prev_block_hash: &'a CryptoHash,
@@ -93,10 +67,6 @@ impl<'a> RuntimeExt<'a> {
         RuntimeExt {
             trie_update,
             account_id,
-            action_receipts: vec![],
-            signer_id,
-            signer_public_key,
-            gas_price,
             action_hash,
             data_count: 0,
             epoch_id,
@@ -104,9 +74,8 @@ impl<'a> RuntimeExt<'a> {
             last_block_hash,
             epoch_info_provider,
             current_protocol_version,
-
-            #[cfg(feature = "protocol_feature_function_call_weight")]
-            gas_weights: vec![],
+            // #[cfg(feature = "protocol_feature_function_call_weight")]
+            // gas_weights: vec![],
         }
     }
 
@@ -128,7 +97,8 @@ impl<'a> RuntimeExt<'a> {
         TrieKey::ContractData { account_id: self.account_id.clone(), key: key.to_vec() }
     }
 
-    fn new_data_id(&mut self) -> CryptoHash {
+    // TODO needs to be moved into trait to be used in VMLogic
+    pub fn new_data_id(&mut self) -> CryptoHash {
         let data_id = create_data_id(
             self.current_protocol_version,
             self.action_hash,
@@ -140,34 +110,34 @@ impl<'a> RuntimeExt<'a> {
         data_id
     }
 
-    pub fn into_receipts(self, predecessor_id: &AccountId) -> Vec<Receipt> {
-        self.action_receipts
-            .into_iter()
-            .map(|(receiver_id, action_receipt)| Receipt {
-                predecessor_id: predecessor_id.clone(),
-                receiver_id,
-                // Actual receipt ID is set in the Runtime.apply_action_receipt(...) in the
-                // "Generating receipt IDs" section
-                receipt_id: CryptoHash::default(),
-                receipt: ReceiptEnum::Action(action_receipt),
-            })
-            .collect()
-    }
+    // pub fn into_receipts(self, predecessor_id: &AccountId) -> Vec<Receipt> {
+    //     self.action_receipts
+    //         .into_iter()
+    //         .map(|(receiver_id, action_receipt)| Receipt {
+    //             predecessor_id: predecessor_id.clone(),
+    //             receiver_id,
+    //             // Actual receipt ID is set in the Runtime.apply_action_receipt(...) in the
+    //             // "Generating receipt IDs" section
+    //             receipt_id: CryptoHash::default(),
+    //             receipt: ReceiptEnum::Action(action_receipt),
+    //         })
+    //         .collect()
+    // }
 
-    /// Appends an action and returns the index the action was inserted in the receipt
-    fn append_action(&mut self, receipt_index: u64, action: Action) -> usize {
-        let actions = &mut self
-            .action_receipts
-            .get_mut(receipt_index as usize)
-            .expect("receipt index should be present")
-            .1
-            .actions;
+    // /// Appends an action and returns the index the action was inserted in the receipt
+    // fn append_action(&mut self, receipt_index: u64, action: Action) -> usize {
+    //     let actions = &mut self
+    //         .action_receipts
+    //         .get_mut(receipt_index as usize)
+    //         .expect("receipt index should be present")
+    //         .1
+    //         .actions;
 
-        actions.push(action);
+    //     actions.push(action);
 
-        // Return index that action was inserted at
-        actions.len() - 1
-    }
+    //     // Return index that action was inserted at
+    //     actions.len() - 1
+    // }
 
     pub fn set_trie_cache_mode(&mut self, state: TrieCacheMode) {
         self.trie_update.set_trie_cache_mode(state);
@@ -234,200 +204,200 @@ impl<'a> External for RuntimeExt<'a> {
         Ok(())
     }
 
-    fn create_receipt(
-        &mut self,
-        receipt_indices: Vec<u64>,
-        receiver_id: AccountId,
-    ) -> ExtResult<u64> {
-        let mut input_data_ids = vec![];
-        for receipt_index in receipt_indices {
-            let data_id = self.new_data_id();
-            self.action_receipts
-                .get_mut(receipt_index as usize)
-                .ok_or_else(|| HostError::InvalidReceiptIndex { receipt_index })?
-                .1
-                .output_data_receivers
-                .push(DataReceiver { data_id, receiver_id: receiver_id.clone() });
-            input_data_ids.push(data_id);
-        }
+    // fn create_receipt(
+    //     &mut self,
+    //     receipt_indices: Vec<u64>,
+    //     receiver_id: AccountId,
+    // ) -> ExtResult<u64> {
+    //     let mut input_data_ids = vec![];
+    //     for receipt_index in receipt_indices {
+    //         let data_id = self.new_data_id();
+    //         self.action_receipts
+    //             .get_mut(receipt_index as usize)
+    //             .ok_or_else(|| HostError::InvalidReceiptIndex { receipt_index })?
+    //             .1
+    //             .output_data_receivers
+    //             .push(DataReceiver { data_id, receiver_id: receiver_id.clone() });
+    //         input_data_ids.push(data_id);
+    //     }
 
-        let new_receipt = ActionReceipt {
-            signer_id: self.signer_id.clone(),
-            signer_public_key: self.signer_public_key.clone(),
-            gas_price: self.gas_price,
-            output_data_receivers: vec![],
-            input_data_ids,
-            actions: vec![],
-        };
-        let new_receipt_index = self.action_receipts.len() as u64;
-        self.action_receipts.push((receiver_id, new_receipt));
-        Ok(new_receipt_index)
-    }
+    //     let new_receipt = ActionReceipt {
+    //         signer_id: self.signer_id.clone(),
+    //         signer_public_key: self.signer_public_key.clone(),
+    //         gas_price: self.gas_price,
+    //         output_data_receivers: vec![],
+    //         input_data_ids,
+    //         actions: vec![],
+    //     };
+    //     let new_receipt_index = self.action_receipts.len() as u64;
+    //     self.action_receipts.push((receiver_id, new_receipt));
+    //     Ok(new_receipt_index)
+    // }
 
-    fn append_action_create_account(&mut self, receipt_index: u64) -> ExtResult<()> {
-        self.append_action(receipt_index, Action::CreateAccount(CreateAccountAction {}));
-        Ok(())
-    }
+    // fn append_action_create_account(&mut self, receipt_index: u64) -> ExtResult<()> {
+    //     self.append_action(receipt_index, Action::CreateAccount(CreateAccountAction {}));
+    //     Ok(())
+    // }
 
-    fn append_action_deploy_contract(
-        &mut self,
-        receipt_index: u64,
-        code: Vec<u8>,
-    ) -> ExtResult<()> {
-        self.append_action(receipt_index, Action::DeployContract(DeployContractAction { code }));
-        Ok(())
-    }
+    // fn append_action_deploy_contract(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     code: Vec<u8>,
+    // ) -> ExtResult<()> {
+    //     self.append_action(receipt_index, Action::DeployContract(DeployContractAction { code }));
+    //     Ok(())
+    // }
 
-    #[cfg(feature = "protocol_feature_function_call_weight")]
-    fn append_action_function_call_weight(
-        &mut self,
-        receipt_index: u64,
-        method_name: Vec<u8>,
-        args: Vec<u8>,
-        attached_deposit: u128,
-        prepaid_gas: Gas,
-        gas_weight: GasWeight,
-    ) -> ExtResult<()> {
-        let action_index = self.append_action(
-            receipt_index,
-            Action::FunctionCall(FunctionCallAction {
-                method_name: String::from_utf8(method_name)
-                    .map_err(|_| HostError::InvalidMethodName)?,
-                args,
-                gas: prepaid_gas,
-                deposit: attached_deposit,
-            }),
-        );
+    // #[cfg(feature = "protocol_feature_function_call_weight")]
+    // fn append_action_function_call_weight(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     method_name: Vec<u8>,
+    //     args: Vec<u8>,
+    //     attached_deposit: u128,
+    //     prepaid_gas: Gas,
+    //     gas_weight: GasWeight,
+    // ) -> ExtResult<()> {
+    //     let action_index = self.append_action(
+    //         receipt_index,
+    //         Action::FunctionCall(FunctionCallAction {
+    //             method_name: String::from_utf8(method_name)
+    //                 .map_err(|_| HostError::InvalidMethodName)?,
+    //             args,
+    //             gas: prepaid_gas,
+    //             deposit: attached_deposit,
+    //         }),
+    //     );
 
-        if gas_weight.0 > 0 {
-            self.gas_weights.push((
-                FunctionCallActionIndex { receipt_index: receipt_index as usize, action_index },
-                gas_weight,
-            ));
-        }
+    //     if gas_weight.0 > 0 {
+    //         self.gas_weights.push((
+    //             FunctionCallActionIndex { receipt_index: receipt_index as usize, action_index },
+    //             gas_weight,
+    //         ));
+    //     }
 
-        Ok(())
-    }
+    //     Ok(())
+    // }
 
-    fn append_action_function_call(
-        &mut self,
-        receipt_index: u64,
-        method_name: Vec<u8>,
-        args: Vec<u8>,
-        attached_deposit: u128,
-        prepaid_gas: Gas,
-    ) -> ExtResult<()> {
-        self.append_action(
-            receipt_index,
-            Action::FunctionCall(FunctionCallAction {
-                method_name: String::from_utf8(method_name)
-                    .map_err(|_| HostError::InvalidMethodName)?,
-                args,
-                gas: prepaid_gas,
-                deposit: attached_deposit,
-            }),
-        );
-        Ok(())
-    }
+    // fn append_action_function_call(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     method_name: Vec<u8>,
+    //     args: Vec<u8>,
+    //     attached_deposit: u128,
+    //     prepaid_gas: Gas,
+    // ) -> ExtResult<()> {
+    //     self.append_action(
+    //         receipt_index,
+    //         Action::FunctionCall(FunctionCallAction {
+    //             method_name: String::from_utf8(method_name)
+    //                 .map_err(|_| HostError::InvalidMethodName)?,
+    //             args,
+    //             gas: prepaid_gas,
+    //             deposit: attached_deposit,
+    //         }),
+    //     );
+    //     Ok(())
+    // }
 
-    fn append_action_transfer(&mut self, receipt_index: u64, deposit: u128) -> ExtResult<()> {
-        self.append_action(receipt_index, Action::Transfer(TransferAction { deposit }));
-        Ok(())
-    }
+    // fn append_action_transfer(&mut self, receipt_index: u64, deposit: u128) -> ExtResult<()> {
+    //     self.append_action(receipt_index, Action::Transfer(TransferAction { deposit }));
+    //     Ok(())
+    // }
 
-    fn append_action_stake(
-        &mut self,
-        receipt_index: u64,
-        stake: u128,
-        public_key: Vec<u8>,
-    ) -> ExtResult<()> {
-        self.append_action(
-            receipt_index,
-            Action::Stake(StakeAction {
-                stake,
-                public_key: PublicKey::try_from_slice(&public_key)
-                    .map_err(|_| HostError::InvalidPublicKey)?,
-            }),
-        );
-        Ok(())
-    }
+    // fn append_action_stake(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     stake: u128,
+    //     public_key: Vec<u8>,
+    // ) -> ExtResult<()> {
+    //     self.append_action(
+    //         receipt_index,
+    //         Action::Stake(StakeAction {
+    //             stake,
+    //             public_key: PublicKey::try_from_slice(&public_key)
+    //                 .map_err(|_| HostError::InvalidPublicKey)?,
+    //         }),
+    //     );
+    //     Ok(())
+    // }
 
-    fn append_action_add_key_with_full_access(
-        &mut self,
-        receipt_index: u64,
-        public_key: Vec<u8>,
-        nonce: u64,
-    ) -> ExtResult<()> {
-        self.append_action(
-            receipt_index,
-            Action::AddKey(AddKeyAction {
-                public_key: PublicKey::try_from_slice(&public_key)
-                    .map_err(|_| HostError::InvalidPublicKey)?,
-                access_key: AccessKey { nonce, permission: AccessKeyPermission::FullAccess },
-            }),
-        );
-        Ok(())
-    }
+    // fn append_action_add_key_with_full_access(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     public_key: Vec<u8>,
+    //     nonce: u64,
+    // ) -> ExtResult<()> {
+    //     self.append_action(
+    //         receipt_index,
+    //         Action::AddKey(AddKeyAction {
+    //             public_key: PublicKey::try_from_slice(&public_key)
+    //                 .map_err(|_| HostError::InvalidPublicKey)?,
+    //             access_key: AccessKey { nonce, permission: AccessKeyPermission::FullAccess },
+    //         }),
+    //     );
+    //     Ok(())
+    // }
 
-    fn append_action_add_key_with_function_call(
-        &mut self,
-        receipt_index: u64,
-        public_key: Vec<u8>,
-        nonce: u64,
-        allowance: Option<u128>,
-        receiver_id: AccountId,
-        method_names: Vec<Vec<u8>>,
-    ) -> ExtResult<()> {
-        self.append_action(
-            receipt_index,
-            Action::AddKey(AddKeyAction {
-                public_key: PublicKey::try_from_slice(&public_key)
-                    .map_err(|_| HostError::InvalidPublicKey)?,
-                access_key: AccessKey {
-                    nonce,
-                    permission: AccessKeyPermission::FunctionCall(FunctionCallPermission {
-                        allowance,
-                        receiver_id: receiver_id.into(),
-                        method_names: method_names
-                            .into_iter()
-                            .map(|method_name| {
-                                String::from_utf8(method_name)
-                                    .map_err(|_| HostError::InvalidMethodName)
-                            })
-                            .collect::<std::result::Result<Vec<_>, _>>()?,
-                    }),
-                },
-            }),
-        );
-        Ok(())
-    }
+    // fn append_action_add_key_with_function_call(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     public_key: Vec<u8>,
+    //     nonce: u64,
+    //     allowance: Option<u128>,
+    //     receiver_id: AccountId,
+    //     method_names: Vec<Vec<u8>>,
+    // ) -> ExtResult<()> {
+    //     self.append_action(
+    //         receipt_index,
+    //         Action::AddKey(AddKeyAction {
+    //             public_key: PublicKey::try_from_slice(&public_key)
+    //                 .map_err(|_| HostError::InvalidPublicKey)?,
+    //             access_key: AccessKey {
+    //                 nonce,
+    //                 permission: AccessKeyPermission::FunctionCall(FunctionCallPermission {
+    //                     allowance,
+    //                     receiver_id: receiver_id.into(),
+    //                     method_names: method_names
+    //                         .into_iter()
+    //                         .map(|method_name| {
+    //                             String::from_utf8(method_name)
+    //                                 .map_err(|_| HostError::InvalidMethodName)
+    //                         })
+    //                         .collect::<std::result::Result<Vec<_>, _>>()?,
+    //                 }),
+    //             },
+    //         }),
+    //     );
+    //     Ok(())
+    // }
 
-    fn append_action_delete_key(
-        &mut self,
-        receipt_index: u64,
-        public_key: Vec<u8>,
-    ) -> ExtResult<()> {
-        self.append_action(
-            receipt_index,
-            Action::DeleteKey(DeleteKeyAction {
-                public_key: PublicKey::try_from_slice(&public_key)
-                    .map_err(|_| HostError::InvalidPublicKey)?,
-            }),
-        );
-        Ok(())
-    }
+    // fn append_action_delete_key(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     public_key: Vec<u8>,
+    // ) -> ExtResult<()> {
+    //     self.append_action(
+    //         receipt_index,
+    //         Action::DeleteKey(DeleteKeyAction {
+    //             public_key: PublicKey::try_from_slice(&public_key)
+    //                 .map_err(|_| HostError::InvalidPublicKey)?,
+    //         }),
+    //     );
+    //     Ok(())
+    // }
 
-    fn append_action_delete_account(
-        &mut self,
-        receipt_index: u64,
-        beneficiary_id: AccountId,
-    ) -> ExtResult<()> {
-        self.append_action(
-            receipt_index,
-            Action::DeleteAccount(DeleteAccountAction { beneficiary_id }),
-        );
-        Ok(())
-    }
+    // fn append_action_delete_account(
+    //     &mut self,
+    //     receipt_index: u64,
+    //     beneficiary_id: AccountId,
+    // ) -> ExtResult<()> {
+    //     self.append_action(
+    //         receipt_index,
+    //         Action::DeleteAccount(DeleteAccountAction { beneficiary_id }),
+    //     );
+    //     Ok(())
+    // }
 
     fn get_touched_nodes_count(&self) -> u64 {
         self.trie_update.trie.get_touched_nodes_count()
@@ -445,54 +415,54 @@ impl<'a> External for RuntimeExt<'a> {
             .map_err(|e| ExternalError::ValidatorError(e).into())
     }
 
-    /// Distributes the gas passed in by splitting it among weights defined in `gas_weights`.
-    /// This will sum all weights, retrieve the gas per weight, then update each function
-    /// to add the respective amount of gas. Once all gas is distributed, the remainder of
-    /// the gas not assigned due to precision loss is added to the last function with a weight.
-    #[cfg(feature = "protocol_feature_function_call_weight")]
-    fn distribute_unused_gas(&mut self, gas: u64) -> GasDistribution {
-        let gas_weight_sum: u128 =
-            self.gas_weights.iter().map(|(_, GasWeight(weight))| *weight as u128).sum();
-        if gas_weight_sum != 0 {
-            // Floor division that will ensure gas allocated is <= gas to distribute
-            let gas_per_weight = (gas as u128 / gas_weight_sum) as u64;
+    // /// Distributes the gas passed in by splitting it among weights defined in `gas_weights`.
+    // /// This will sum all weights, retrieve the gas per weight, then update each function
+    // /// to add the respective amount of gas. Once all gas is distributed, the remainder of
+    // /// the gas not assigned due to precision loss is added to the last function with a weight.
+    // #[cfg(feature = "protocol_feature_function_call_weight")]
+    // fn distribute_unused_gas(&mut self, gas: u64) -> GasDistribution {
+    //     let gas_weight_sum: u128 =
+    //         self.gas_weights.iter().map(|(_, GasWeight(weight))| *weight as u128).sum();
+    //     if gas_weight_sum != 0 {
+    //         // Floor division that will ensure gas allocated is <= gas to distribute
+    //         let gas_per_weight = (gas as u128 / gas_weight_sum) as u64;
 
-            let mut distribute_gas = |metadata: &FunctionCallActionIndex, assigned_gas: u64| {
-                let FunctionCallActionIndex { receipt_index, action_index } = metadata;
-                if let Some(Action::FunctionCall(FunctionCallAction { ref mut gas, .. })) = self
-                    .action_receipts
-                    .get_mut(*receipt_index)
-                    .and_then(|(_, receipt)| receipt.actions.get_mut(*action_index))
-                {
-                    *gas += assigned_gas;
-                } else {
-                    panic!(
-                        "Invalid index for assigning unused gas weight \
-                        (promise_index={}, action_index={})",
-                        receipt_index, action_index
-                    );
-                }
-            };
+    //         let mut distribute_gas = |metadata: &FunctionCallActionIndex, assigned_gas: u64| {
+    //             let FunctionCallActionIndex { receipt_index, action_index } = metadata;
+    //             if let Some(Action::FunctionCall(FunctionCallAction { ref mut gas, .. })) = self
+    //                 .action_receipts
+    //                 .get_mut(*receipt_index)
+    //                 .and_then(|(_, receipt)| receipt.actions.get_mut(*action_index))
+    //             {
+    //                 *gas += assigned_gas;
+    //             } else {
+    //                 panic!(
+    //                     "Invalid index for assigning unused gas weight \
+    //                     (promise_index={}, action_index={})",
+    //                     receipt_index, action_index
+    //                 );
+    //             }
+    //         };
 
-            let mut distributed = 0;
-            for (action_index, GasWeight(weight)) in &self.gas_weights {
-                // This can't overflow because the gas_per_weight is floor division
-                // of the weight sum.
-                let assigned_gas = gas_per_weight * weight;
+    //         let mut distributed = 0;
+    //         for (action_index, GasWeight(weight)) in &self.gas_weights {
+    //             // This can't overflow because the gas_per_weight is floor division
+    //             // of the weight sum.
+    //             let assigned_gas = gas_per_weight * weight;
 
-                distribute_gas(action_index, assigned_gas);
+    //             distribute_gas(action_index, assigned_gas);
 
-                distributed += assigned_gas
-            }
+    //             distributed += assigned_gas
+    //         }
 
-            // Distribute remaining gas to final action.
-            if let Some((last_idx, _)) = self.gas_weights.last() {
-                distribute_gas(last_idx, gas - distributed);
-            }
-            self.gas_weights.clear();
-            GasDistribution::All
-        } else {
-            GasDistribution::NoRatios
-        }
-    }
+    //         // Distribute remaining gas to final action.
+    //         if let Some((last_idx, _)) = self.gas_weights.last() {
+    //             distribute_gas(last_idx, gas - distributed);
+    //         }
+    //         self.gas_weights.clear();
+    //         GasDistribution::All
+    //     } else {
+    //         GasDistribution::NoRatios
+    //     }
+    // }
 }
